[WIP] SPARKNLP-831: Add TypeConversion to dynamic setters
